Decoders: Move UnionFindDecoder to hdrg_decoders.py, rename ClusteringDecoder to BravyiHaahDecoder and create generic base class for clustering decoders
