+++ qiskit_qec/decoders/__init__.py
@@ -34,4 +34,4 @@ from .decoding_graph import DecodingGraph
 from .circuit_matching_decoder import CircuitModelMatchingDecoder
 from .repetition_decoder import RepetitionDecoder
 from .three_bit_decoder import ThreeBitDecoder
-from .union_find import UnionFindDecoder
+from .hdrg_decoders import BravyiHaahDecoder, UnionFindDecoder
